Convert project to julia lang
Julia is such a nice language. This commit ports the fractus
functionalities to it. The julia script creates an api which
can generate the requested fractal as a png. Right now we save
that file locally, but this can be adapted to return it to the
requester. Anyway, damn, this turned out nice. New favorite
language without a doubt.